move the `BaseSimulationData`, `BaseRiverData`, `LineGeometry`, `_read_fm_map` to a `io.data_models` module.
