+++ src/dfastbe/bank_erosion/data_models.py
@@ -8,7 +8,8 @@ from geopandas import GeoDataFrame
 from shapely.geometry import LineString, Point
 from geopandas.geoseries import GeoSeries
 from dfastio.xyc.models import XYCModel
-from dfastbe.io.io import ConfigFile, BaseRiverData, BaseSimulationData, log_text
+from dfastbe.io.data_models import BaseRiverData, BaseSimulationData
+from dfastbe.io.io import ConfigFile, log_text
 
 
 GenericType = TypeVar("GenericType")
