+++ src/dfastbe/bank_lines/bank_lines.py
@@ -14,7 +14,8 @@ from shapely.geometry.polygon import Polygon
 from dfastbe import __version__
 from dfastbe import plotting as df_plt
 from dfastbe.bank_lines.data_models import BankLinesRiverData
-from dfastbe.io.io import BaseSimulationData, ConfigFile, LineGeometry, get_bbox, log_text
+from dfastbe.io.data_models import BaseSimulationData, LineGeometry
+from dfastbe.io.io import ConfigFile, get_bbox, log_text
 from dfastbe.kernel import get_zoom_extends
 from dfastbe.support import (
     on_right_side,
