+++ tests/test_io.py
@@ -13,20 +13,16 @@ import pytest
 from geopandas import GeoDataFrame
 from pyfakefs.fake_filesystem import FakeFilesystem
 from shapely.geometry import LineString
-
+from dfastbe.io.data_models import LineGeometry, BaseSimulationData, BaseRiverData, _read_fm_map
 from dfastbe.io.io import (
-    LineGeometry,
-    BaseSimulationData,
     SimulationFilesError,
     ConfigFile,
-    BaseRiverData,
     absolute_path,
     get_filename,
     get_text,
     load_program_texts,
-    log_text,
-    _read_fm_map,
     relative_path,
+    log_text
 )
 
 
@@ -67,7 +63,7 @@ class TestSimulationData:
         mock_velocity_y_face = np.array([0.4, 0.5, 0.6])
         mock_chezy_face = np.array([30.0, 40.0, 50.0])
 
-        with patch("dfastbe.io.io._read_fm_map") as mock_read_fm_map, patch(
+        with patch("dfastbe.io.data_models._read_fm_map") as mock_read_fm_map, patch(
             "netCDF4.Dataset"
         ) as mock_dataset:
             mock_read_fm_map.side_effect = [
